refactoring game*.ts files
